requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SocketIO==2.9.3
havenondemand==2.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
python-decouple==3.1
python-engineio==2.0.1
python-socketio==1.8.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
tweepy==3.5.0
twitter==1.18.0
urllib3==1.22
Werkzeug==0.13
